(release) version 5.16

+++ app/src/main/java/com/artmaster/android/orthodoxcalendar/ui/RoutersComponent.kt
@@ -71,8 +71,7 @@ fun AppNavigationComponent(
         startDestination = startRoute
     ) {
         composable(Navigation.INIT_PAGE.route) {
-            AppStartTextAnimation(duration = initViewModel.animationTime.toInt(), resIndex = null) {
-
+            AppStartTextAnimation(duration = initViewModel.animationTime, resIndex = null) {
                 val route = if (calendarViewModel.firstLoadingTileCalendar())
                     Navigation.TILE_CALENDAR
                 else Navigation.LIST_CALENDAR
